requirements.txt
streamlit
langchain
openai
python-dotenv
sqlalchemy
pandas
streamlit-chat
matplotlib
requests
azure-openai
langchain-openai
pyowm